refactor: consolidate homepage components and enhance layout
- Merged homepage layout and page components into a single HomePage component for improved structure and readability.
- Removed the previous layout file and updated the main page to include a centered layout with a title, description, and action links for user engagement.
